move create room kind selector to components
